documentation, rename "subsystems"->"gamepiece"
