+++ src/main/java/frc4146/robot/commands/gamepiece/ClawCommand.java
@@ -1,4 +1,4 @@
-package frc4146.robot.commands.subsystems;
+package frc4146.robot.commands.gamepiece;
 
 import common.robot.input.Axis;
 import edu.wpi.first.wpilibj2.command.CommandBase;
