+++ src/main/java/frc4146/robot/commands/gamepiece/PositionPiece.java
@@ -1,4 +1,4 @@
-package frc4146.robot.commands.subsystems;
+package frc4146.robot.commands.gamepiece;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
