+++ src/main/java/frc4146/robot/subsystems/DrivetrainSubsystem.java
@@ -297,7 +297,7 @@ public class DrivetrainSubsystem implements Subsystem, UpdateManager.Updatable {
     updateDisplay();
   }
 
-  /** updates NetworkTable with data read from robot pose */
+  /* updates NetworkTable with data read from robot pose */
   public void updateDisplay() {
     RigidTransform2 pose = getPose();
     odometryXEntry.setDouble(pose.translation.x);
@@ -305,6 +305,7 @@ public class DrivetrainSubsystem implements Subsystem, UpdateManager.Updatable {
     odometryAngleEntry.setDouble(pose.rotation.toDegrees());
   }
 
+  /* allows robot to drive in only one direction while self-leveling */
   public void lockWheelsAngle(double angle) {
     if (getAverageAbsoluteValueVelocity() < 5.0) {
       frontLeftModule.set(0, angle * 2 * Math.PI / 180);
